added BaseMainActivity class
