+++ services_app/src/androidTest/java/org/opendatakit/activites/MainActivity/LoggedOutStateTest.java
@@ -8,7 +8,6 @@ import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
 import static com.google.common.truth.Truth.assertThat;
 
-import android.content.Context;
 import android.content.Intent;
 
 import androidx.test.core.app.ActivityScenario;
@@ -17,9 +16,7 @@ import androidx.test.espresso.intent.Intents;
 import androidx.test.espresso.intent.matcher.IntentMatchers;
 import androidx.test.espresso.matcher.RootMatchers;
 import androidx.test.espresso.matcher.ViewMatchers;
-import androidx.test.platform.app.InstrumentationRegistry;
 
-import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.opendatakit.consts.IntentConsts;
@@ -32,9 +29,7 @@ import org.opendatakit.services.sync.actions.fragments.UpdateServerSettingsFragm
 
 import java.util.Map;
 
-public class LoggedOutStateTest {
-
-    private ActivityScenario<MainActivity> activityScenario;
+public class LoggedOutStateTest extends BaseMainActivity {
 
     @Before
     public void setUp() {
@@ -99,24 +94,4 @@ public class LoggedOutStateTest {
         Intents.intended(IntentMatchers.hasComponent(LoginActivity.class.getName()));
         Intents.release();
     }
-
-    @After
-    public void clearTestEnvironment() {
-        activityScenario.onActivity(activity -> {
-            PropertiesSingleton props = CommonToolProperties.get(activity, activity.getAppName());
-            assertThat(props).isNotNull();
-
-            Map<String, String> serverProperties = UpdateServerSettingsFragment.getUpdateUrlProperties(
-                    activity.getString(org.opendatakit.androidlibrary.R.string.default_sync_server_url)
-            );
-            assertThat(serverProperties).isNotNull();
-            serverProperties.put(CommonToolProperties.KEY_FIRST_LAUNCH, "true");
-            props.setProperties(serverProperties);
-        });
-    }
-
-    private Context getContext() {
-        return InstrumentationRegistry.getInstrumentation().getTargetContext();
-    }
-
 }
